removed default for preblast time, started adding testing, handler code cleanup
